refactor(#6) - 피드백 반영
- 부적절한 메서드 네이밍 수정
- DiscountValue의 타입을 BigDeicmal -> int로 변경
- 주문 생성 시 쿠폰을 확인하고 반환할 때 Optional 타입으로 반환

+++ src/test/java/com/flab/infrun/order/application/CreateOrderProcessorTest.java
@@ -125,7 +125,7 @@ final class CreateOrderProcessorTest extends IntegrationTest {
         couponRepository.save(
             aCouponFixture()
                 .discountInfo(
-                    DiscountInfo.of(DiscountType.FIX, BigDecimal.valueOf(1_000)))
+                    DiscountInfo.of(DiscountType.FIX, 1_000))
                 .build());
     }
 }
